Extract middleware from service files
