+++ src/api/services/participantsService.ts
@@ -1,4 +1,4 @@
-import { NextFunction, Request, Response } from 'express';
+import { Request } from 'express';
 import { TransactionOrKnex } from 'objection';
 import { z } from 'zod';
 
@@ -24,7 +24,6 @@ import {
 } from './auditTrailService';
 import { createEmailService } from './emailService';
 import { EmailArgs } from './emailTypes';
-import { findUserByEmail, isUserBelongsToParticipant } from './usersService';
 
 export interface ParticipantRequest extends Request {
   participant?: Participant;
@@ -301,60 +300,4 @@ export const sendParticipantApprovedEmail = async (users: User[], traceId: strin
     to: users.map((user) => ({ name: user.fullName(), email: user.email })),
   };
   emailService.sendEmail(emailArgs, traceId);
-};
-
-const idParser = z.object({
-  participantId: z.coerce.number(),
-});
-
-// TODO: move this middleware to a separate file
-const hasParticipantAccess = async (req: ParticipantRequest, res: Response, next: NextFunction) => {
-  const { participantId } = idParser.parse(req.params);
-  const traceId = getTraceId(req);
-  const participant = await Participant.query().findById(participantId).withGraphFetched('types');
-  if (!participant) {
-    return res.status(404).send([{ message: 'The participant cannot be found.' }]);
-  }
-
-  const currentUserEmail = req.auth?.payload?.email as string;
-  if (!(await isUserBelongsToParticipant(currentUserEmail, participantId, traceId))) {
-    return res.status(403).send([{ message: 'You do not have permission to that participant.' }]);
-  }
-
-  req.participant = participant;
-  return next();
-};
-
-// TODO: move this middleware to a separate file
-const enrichCurrentParticipant = async (
-  req: ParticipantRequest,
-  res: Response,
-  next: NextFunction
-) => {
-  const userEmail = req.auth?.payload?.email as string;
-  const user = await findUserByEmail(userEmail);
-  if (!user) {
-    return res.status(404).send([{ message: 'The user cannot be found.' }]);
-  }
-  // TODO: This just gets the user's first participant, but it will need to get the currently selected participant as part of UID2-2822
-  const participant = user.participants?.[0];
-
-  if (!participant) {
-    return res.status(404).send([{ message: 'The participant cannot be found.' }]);
-  }
-  req.participant = participant;
-  return next();
-};
-
-// TODO: move this middleware to a separate file
-export const checkParticipantId = async (
-  req: ParticipantRequest,
-  res: Response,
-  next: NextFunction
-) => {
-  // TODO: Remove support for 'current' in UID2-2822
-  if (req.params.participantId === 'current') {
-    return enrichCurrentParticipant(req, res, next);
-  }
-  return hasParticipantAccess(req, res, next);
 };
